change a code format

+++ src/t_hash.c
@@ -2214,7 +2214,7 @@ static int parseHashFieldExpireArgs(client *c, int *flags,
             *field_count = (int) val;
 
             /* Validate field count based on command type */
-            int required_args = *first_field_pos + (*field_count * args_per_field);
+            long long required_args = *first_field_pos + ((long long)*field_count * args_per_field);
             if (required_args > c->argc) {
                 addReplyError(c, "wrong number of arguments");
                 return C_ERR;
@@ -2233,7 +2233,6 @@ static int parseHashFieldExpireArgs(client *c, int *flags,
 
             *flags |= HFE_EX;
             i++;
-
             if (parseExpireTime(c, c->argv[i], UNIT_SECONDS,
                                 commandTimeSnapshot(), expire_time) != C_OK)
                 return C_ERR;
@@ -2248,7 +2247,6 @@ static int parseHashFieldExpireArgs(client *c, int *flags,
 
             *flags |= HFE_PX;
             i++;
-
             if (parseExpireTime(c, c->argv[i], UNIT_MILLISECONDS,
                                 commandTimeSnapshot(), expire_time) != C_OK)
                 return C_ERR;
@@ -2263,7 +2261,6 @@ static int parseHashFieldExpireArgs(client *c, int *flags,
 
             *flags |= HFE_EXAT;
             i++;
-
             if (parseExpireTime(c, c->argv[i], UNIT_SECONDS, 0, expire_time) != C_OK)
                 return C_ERR;
 
@@ -2277,8 +2274,8 @@ static int parseHashFieldExpireArgs(client *c, int *flags,
 
             *flags |= HFE_PXAT;
             i++;
-
-            if (parseExpireTime(c, c->argv[i], UNIT_MILLISECONDS, 0, expire_time) != C_OK)
+            if (parseExpireTime(c, c->argv[i], UNIT_MILLISECONDS, 0,
+                                expire_time) != C_OK)
                 return C_ERR;
 
             *expire_time_pos = i;
